pyproject.toml
[tool.poetry]
name = "llm-routing"
version = "0.1.0"
description = ""
authors = ["lemberck <blemberck@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.47.0"
python-dotenv = "^1.0.1"
termcolor = "^2.4.0"
timeout-decorator = "^0.5.0"
langchain-openai = "^0.2.0"
langchain-core = "^0.3.5"
langgraph = "^0.2.23"
langchain-community = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
